requirements.txt
django
django-cors-headers
django-rest-framework
djangorestframework-simplejwt
